// frontend/src/payload-types.ts
export interface SiteSetting {
  id: number;
  tagline?: string | null;
  announcement?: string | null;
  supportEmail?: string | null;
  supportPhone?: string | null;
  social?: {
    twitter?: string | null;
    instagram?: string | null;
    tiktok?: string | null;
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}